move acurl utils types to module
